Fixed undefined issue

+++ tests/UnitTests/UnitTest_AVX512/CodeDeprecated-avx512.cc
@@ -237,17 +237,18 @@ int dwt_horizontal_depricated_avx512(int32_t* out_lf, int32_t* out_hf, const int
             }
 
             if (remaining > 1 && simd_batch > 0) {
-                dst_lf_row[0] = src_row[0] + ((dst_hf_row[-1] + dst_hf_row[0] + 2) >> 2);
+                // Widen to int64 so the intermediate sums cannot overflow int32 (UBSan).
+                dst_lf_row[0] = (int32_t)((int64_t)src_row[0] + (((int64_t)dst_hf_row[-1] + dst_hf_row[0] + 2) >> 2));
             }
             else if (remaining > 1) {
                 // Fix: When simd_batch==0, dst_hf_row[-1] is before the allocated buffer.
                 // Use the boundary formula instead; this value gets overwritten by
                 // "correct the first for lf" anyway.
-                dst_lf_row[0] = src_row[0] + ((dst_hf_row[0] + 1) >> 1);
+                dst_lf_row[0] = (int32_t)((int64_t)src_row[0] + (((int64_t)dst_hf_row[0] + 1) >> 1));
             }
 
             for (uint32_t i = 1; i < (remaining / 2); i++) {
-                dst_lf_row[i] = src_row[2 * i] + ((dst_hf_row[i - 1] + dst_hf_row[i] + 2) >> 2);
+                dst_lf_row[i] = (int32_t)((int64_t)src_row[2 * i] + (((int64_t)dst_hf_row[i - 1] + dst_hf_row[i] + 2) >> 2));
             }
             dst_lf_row += remaining / 2;
             dst_hf_row += remaining / 2;
@@ -256,7 +257,7 @@ int dwt_horizontal_depricated_avx512(int32_t* out_lf, int32_t* out_hf, const int
 
         // correct the last for hf and lf
         if (width % 2) {
-            dst_lf_row[0] = src_row[-1] + ((dst_hf_row[-1] + 1) >> 1);
+            dst_lf_row[0] = (int32_t)((int64_t)src_row[-1] + (((int64_t)dst_hf_row[-1] + 1) >> 1));
         }
         else {
             dst_hf_row[-1] = src_row[-1] - src_row[-2];
@@ -264,7 +265,7 @@ int dwt_horizontal_depricated_avx512(int32_t* out_lf, int32_t* out_hf, const int
             // the buffer. Skip the LF correction here; "correct the first for lf" below
             // will write the correct value for LF[0].
             if (width >= 4) {
-                dst_lf_row[-1] = src_row[-2] + ((dst_hf_row[-2] + dst_hf_row[-1] + 2) >> 2);
+                dst_lf_row[-1] = (int32_t)((int64_t)src_row[-2] + (((int64_t)dst_hf_row[-2] + dst_hf_row[-1] + 2) >> 2));
             }
         }
